% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annual_cpue}
\alias{annual_cpue}
\title{Annual CPUE indices by group}
\format{
A data.frame with 47 rows and 4 variables:
\describe{
  \item{year}{year.}
  \item{Cymbium}{CPUE \emph{Cymbium} spp.}
  \item{Octopus}{CPUE \emph{Octopus vulgaris}.}
  \item{Penaeus}{CPUE \emph{Penaeus notialis}.}
}
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
annual_cpue
}
\description{
Annual catch per unit effort (abundance index) for the
three target groups.
}
\keyword{datasets}
