% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_quality_score}
\alias{calculate_quality_score}
\title{Calculate a quality score}
\usage{
calculate_quality_score(checks)
}
\arguments{
\item{checks}{list of checks}
}
\value{
score between 0 and 100
}
\description{
Calculate a quality score
}
