% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{check_lfq}
\alias{check_lfq}
\title{Verify the validity of an LFQ object}
\usage{
check_lfq(lfq)
}
\arguments{
\item{lfq}{Object to verify.}
}
\value{
\code{TRUE} invisibly if the object is valid ; raises an
  error otherwise.
}
\description{
Checks that an object is indeed of class \code{lfq} (TropFishR) and that it
contains the required fields (\code{dates}, \code{midLengths}, \code{catch}).
}
\examples{
\dontrun{
  check_lfq(lfq)
}
}
\seealso{
\code{\link{prepare_tropfish}}, \code{\link{summary_lfq}}
}
