% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_required_columns}
\alias{check_required_columns}
\title{Check required columns}
\usage{
check_required_columns(data, required)
}
\arguments{
\item{data}{data.frame}

\item{required}{vector of names}
}
\value{
list
}
\description{
Check required columns
}
