% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{compare_M_methods}
\alias{compare_M_methods}
\title{Comparison of M methods with associated F, E and F/M}
\usage{
compare_M_methods(Z, M_obj, E_target = 0.5)
}
\arguments{
\item{Z}{Total mortality (from \code{estimate_catchcurve}).}

\item{M_obj}{\code{FishStockM} object (from \code{estimate_M_all}).}

\item{E_target}{Exploitation reference.}
}
\value{
Sorted data.frame, one row per M method.
}
\description{
Comparison of M methods with associated F, E and F/M
}
