% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cymbium_bio}
\alias{cymbium_bio}
\title{Individual biological data - Cymbium spp.}
\format{
A data.frame with 3130 rows and 10 variables:
\describe{
  \item{landing_site}{Landing site (factor).}
  \item{Date}{Sampling date (Date).}
  \item{fishing_zone}{Declared fishing ground (factor).}
  \item{gear}{Fishing gear (factor).}
  \item{season}{Hydrological season (factor).}
  \item{species}{\emph{Cymbium} species (factor, 4 levels).}
  \item{LCQ}{Shell length (cm).}
  \item{weight}{Individual weight (g).}
  \item{sex}{Sex: M / F (factor).}
  \item{maturity}{Maturity stage: IM (immature), M (mature),
    P (spawning) (factor).}
}
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
cymbium_bio
}
\description{
Individual measurements (size, weight), sex and maturity stage
simulated for landings of \emph{Cymbium} spp. (fictitious data).
}
\keyword{datasets}
