% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cymbium_freq}
\alias{cymbium_freq}
\title{Size frequencies - Cymbium spp.}
\format{
A data.frame with 8243 rows and 9 variables (same descriptive
  columns as \code{\link{cymbium_bio}}, with \code{LCQ} for the shell
  length and \code{sex}).
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
cymbium_freq
}
\description{
Size measurements from the sampling of catches of
\emph{Cymbium} spp. (input structure for LFQ analyses).
}
\seealso{
\code{\link{prepare_tropfish}} to build an lfq object.
}
\keyword{datasets}
