% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{detect_file_type}
\alias{detect_file_type}
\title{Importing fisheries data}
\usage{
detect_file_type(file)
}
\arguments{
\item{file}{file path}
}
\value{
file extension
}
\description{
This module allows reading, cleaning and standardizing
the data required for analyses:
}
\details{
- length frequencies
- maturity
- catches
- effort
- parameters

Detect the format of a file
}
\keyword{internal}
