% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{extract_growth_parameters}
\alias{extract_growth_parameters}
\title{Extract the growth parameters of an ELEFAN model}
\usage{
extract_growth_parameters(model)
}
\arguments{
\item{model}{Object \code{FishStockELEFAN} (see \code{\link{run_elefan_ga}}).}
}
\value{
A named list : \code{Linf}, \code{K}, \code{t_anchor},
  \code{phiL}, \code{Rn_max}, \code{ASP}, \code{fESP}, \code{elapsed}.
}
\description{
Robustly retrieves \eqn{L_{inf}}, \eqn{K}, \code{t_anchor}, the performance
index \eqn{\phi'} as well as the fit scores
(\code{Rn_max}, \code{ASP}, \code{fESP}) and the computation time.
}
\details{
\eqn{\phi'} is recomputed if missing : \eqn{\phi' = \log_{10}(K) +
2\log_{10}(L_{inf})}.
}
\examples{
\dontrun{
  p <- extract_growth_parameters(fit)
  p$Linf; p$K; p$phiL
}
}
\references{
Pauly, D., Munro, J.L. (1984). Once more on the comparison of
  growth in fish and invertebrates. \emph{ICLARM Fishbyte} 2(1).
}
\seealso{
\code{\link{growth_table}}, \code{\link{compare_growth_models}}
}
