% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbb.R
\name{extract_lbb}
\alias{extract_lbb}
\title{Extraction of annual indicators from an LBB output}
\usage{
extract_lbb(fit)
}
\arguments{
\item{fit}{Object returned by the LBB function.}
}
\value{
data.frame: Year, Linf, Lc, Lopt, MK, FK, FM, ZK, BB0, BBmsy,
  LmeanLopt (columns available depending on the output).
}
\description{
Robustly searches for the key indicators in the object returned by
the \code{LBB} package and assembles them into an annual data.frame.
}
