% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbspr.R
\name{extract_lbspr}
\alias{extract_lbspr}
\title{Extraction of yearly LBSPR estimates}
\usage{
extract_lbspr(fit, spr_target = 0.4, spr_limit = 0.2)
}
\arguments{
\item{fit}{Object \code{LB_obj} returned by \code{LBSPR::LBSPRfit()}.}

\item{spr_target, spr_limit}{Targets for the status diagnostic.}
}
\value{
data.frame: Year, SL50, SL95, FM, SPR, status.
}
\description{
Extraction of yearly LBSPR estimates
}
