% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth_table}
\alias{growth_table}
\title{Table of parameters of a growth model}
\usage{
growth_table(model)
}
\arguments{
\item{model}{Object \code{FishStockELEFAN}.}
}
\value{
A single-row \code{data.frame} (Linf, K, t_anchor, phiL, Rn_max...).
}
\description{
Formats the estimated parameters of an ELEFAN model as a single-row
\code{data.frame}, convenient for assembling reports.
}
\examples{
\dontrun{
  growth_table(fit)
}
}
\seealso{
\code{\link{extract_growth_parameters}}
}
