% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_stock}
\alias{load_stock}
\title{Load a FishStockData object}
\usage{
load_stock(file = "data/processed/FishStockData.rds")
}
\arguments{
\item{file}{Path of the RDS file.}
}
\value{
The \code{FishStockData} object.
}
\description{
Rereads a \code{FishStockData} object previously saved in RDS format.
}
\examples{
\dontrun{
  stock <- load_stock("data/processed/FishStockData.rds")
}
}
\seealso{
\code{\link{save_stock}}
}
