% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{octopus_bio}
\alias{octopus_bio}
\title{Individual biological data - Octopus vulgaris}
\format{
A data.frame with 1682 rows and 12 variables:
\describe{
  \item{landing_site}{Landing site (factor).}
  \item{Date}{Sampling date (Date).}
  \item{fishing_zone}{Fishing ground (factor).}
  \item{gear}{Fishing gear (factor).}
  \item{season}{Hydrological season (factor).}
  \item{species}{Species (factor).}
  \item{LM}{Mantle length (cm).}
  \item{weight}{Individual weight (g).}
  \item{sex}{Sex: M / F (factor).}
  \item{maturity}{Maturity stage (factor, scale 1-4).}
  \item{total_weight_kg}{Total weight of the sampled catch (kg).}
  \item{sample_weight_kg}{Weight of the measured sample (kg).}
}
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
octopus_bio
}
\description{
Individual measurements, sex and maturity of the common octopus
\emph{Octopus vulgaris} (fictitious data).
}
\keyword{datasets}
