% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{octopus_freq}
\alias{octopus_freq}
\title{Size frequencies - Octopus vulgaris}
\format{
A data.frame with 1689 rows and 9 variables (\code{LM} for the
  mantle length).
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
octopus_freq
}
\description{
Size measurements from the sampling of catches of the octopus
\emph{Octopus vulgaris}.
}
\keyword{datasets}
