% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penaeus_bio}
\alias{penaeus_bio}
\title{Individual biological data - Penaeus notialis}
\format{
A data.frame with 6286 rows and 13 variables:
\describe{
  \item{landing_site}{Landing site (factor).}
  \item{Date}{Sampling date (Date).}
  \item{fishing_zone}{Fishing ground (factor).}
  \item{gear}{Fishing gear (factor).}
  \item{season}{Hydrological season (factor).}
  \item{species}{Species (factor).}
  \item{LCT}{Cephalothoracic length (mm).}
  \item{weight}{Individual weight (g).}
  \item{sex}{Sex: M / F (factor).}
  \item{maturity}{Maturity stage (factor, scale 1-4).}
  \item{gonad_weight}{Gonad weight (g).}
  \item{total_weight_kg}{Total weight of the sampled catch (kg).}
  \item{sample_weight_kg}{Weight of the measured sample (kg).}
}
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
penaeus_bio
}
\description{
Individual measurements, sex, maturity and gonad weight of the
pink shrimp \emph{Penaeus notialis} (fictitious data).
}
\keyword{datasets}
