% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penaeus_freq}
\alias{penaeus_freq}
\title{Size frequencies - Penaeus notialis}
\format{
A data.frame with 31043 rows and 8 variables (\code{LCT} for the
  cephalothoracic length).
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
penaeus_freq
}
\description{
Size measurements from the sampling of catches of the
pink shrimp \emph{Penaeus notialis}.
}
\keyword{datasets}
