% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure.R
\name{plot.FishStockClosure}
\alias{plot.FishStockClosure}
\title{Chart of the monthly closure score with optimal window}
\usage{
\method{plot}{FishStockClosure}(x, ...)
}
\arguments{
\item{x}{\code{FishStockClosure} object.}

\item{...}{Ignored.}
}
\description{
Chart of the monthly closure score with optimal window
}
