% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{plot.FishStockELEFAN}
\alias{plot.FishStockELEFAN}
\title{Plots the response surface of a grid-search ELEFAN fit}
\usage{
\method{plot}{FishStockELEFAN}(x, ...)
}
\arguments{
\item{x}{Object \code{FishStockELEFAN}.}

\item{...}{Arguments passed to \code{\link{plot_response_surface}}.}
}
\value{
A \code{ggplot} object (invisible).
}
\description{
Plots the response surface of a grid-search ELEFAN fit
}
