% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{plot.FishStockKScan}
\alias{plot.FishStockKScan}
\title{Plots a K-Scan (Rn score profile as a function of K)}
\usage{
\method{plot}{FishStockKScan}(x, ...)
}
\arguments{
\item{x}{Object \code{FishStockKScan}.}

\item{...}{Arguments passed to \code{\link{plot_response_surface}}.}
}
\value{
A \code{ggplot} object (invisible).
}
\description{
Plots a K-Scan (Rn score profile as a function of K)
}
