% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbb.R
\name{plot.FishStockLBB}
\alias{plot.FishStockLBB}
\title{Plot of LBB indicators (B/B0, B/Bmsy, F/M)}
\usage{
\method{plot}{FishStockLBB}(x, ...)
}
\arguments{
\item{x}{Object \code{FishStockLBB}.}

\item{...}{Ignored.}
}
\description{
Plot of LBB indicators (B/B0, B/Bmsy, F/M)
}
