% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbspr.R
\name{plot.FishStockLBSPR}
\alias{plot.FishStockLBSPR}
\title{Plot of LBSPR estimates (SPR, SL50/SL95, F/M over time)}
\usage{
\method{plot}{FishStockLBSPR}(x, ...)
}
\arguments{
\item{x}{Object \code{FishStockLBSPR}.}

\item{...}{Ignored.}
}
\description{
Plot of LBSPR estimates (SPR, SL50/SL95, F/M over time)
}
