% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.FishStockLW}
\alias{plot.FishStockLW}
\title{Fitted length-weight relationship plot}
\usage{
\method{plot}{FishStockLW}(x, ...)
}
\arguments{
\item{x}{\code{FishStockLW} object (see \code{\link{fit_length_weight}}).}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Fitted length-weight relationship plot
}
