% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.FishStockM}
\alias{plot.FishStockM}
\title{M estimate comparison plot}
\usage{
\method{plot}{FishStockM}(x, ...)
}
\arguments{
\item{x}{\code{FishStockM} object (see \code{\link{estimate_M_all}}).}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Represents natural mortality estimates by method (points),
sorted, with consensus values (geometric mean, median) as reference.
}
