% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.FishStockMortality}
\alias{plot.FishStockMortality}
\title{Mortality breakdown plot (Z, M, F)}
\usage{
\method{plot}{FishStockMortality}(x, ...)
}
\arguments{
\item{x}{\code{FishStockMortality} object (see \code{\link{run_mortality}}).}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Stacked bars F / M composing Z, with the exploitation rate E annotated.
}
