% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{plot.FishStockMSE}
\alias{plot.FishStockMSE}
\title{Relative yield vs SPR trade-off by scenario}
\usage{
\method{plot}{FishStockMSE}(x, ...)
}
\arguments{
\item{x}{\code{FishStockMSE} object.}

\item{...}{Ignored.}
}
\description{
Relative yield vs SPR trade-off by scenario
}
