% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.FishStockQuota}
\alias{plot.FishStockQuota}
\title{Quota allocation by segment plot}
\usage{
\method{plot}{FishStockQuota}(x, ...)
}
\arguments{
\item{x}{\code{FishStockQuota} object (see \code{\link{allocate_quota}}).}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Bar chart of quotas allocated by segment, sorted.
}
