% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity.R
\name{plot.FishStockSelectivity}
\alias{plot.FishStockSelectivity}
\title{Gear selectivity diagram}
\usage{
\method{plot}{FishStockSelectivity}(x, ...)
}
\arguments{
\item{x}{Object \code{FishStockSelectivity}.}

\item{...}{Ignored.}
}
\description{
Gear selectivity diagram
}
