% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ypr.R
\name{plot.FishStockYPR}
\alias{plot.FishStockYPR}
\title{Yield-per-recruit diagram}
\usage{
\method{plot}{FishStockYPR}(x, ...)
}
\arguments{
\item{x}{\code{FishStockYPR} object.}

\item{...}{Ignored.}
}
\description{
Yield-per-recruit diagram
}
