% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.R
\name{predict.FishStockLW}
\alias{predict.FishStockLW}
\title{weight predicted from length (from a FishStockLW object)}
\usage{
\method{predict}{FishStockLW}(object, length, ...)
}
\arguments{
\item{object}{\code{FishStockLW} object.}

\item{length}{Vector of lengths.}

\item{...}{Ignored.}
}
\value{
Vector of predicted weights.
}
\description{
weight predicted from length (from a FishStockLW object)
}
