% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{print.FishStockData}
\alias{print.FishStockData}
\title{Display of a FishStockData object}
\usage{
\method{print}{FishStockData}(x, ...)
}
\arguments{
\item{x}{FishStockData object}

\item{...}{additional arguments (ignored).}
}
\value{
\code{x}, invisibly (called for its display side effect).
}
\description{
Display of a FishStockData object
}
