% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.FishStockValidation}
\alias{print.FishStockValidation}
\title{Display a FishStockValidation object}
\usage{
\method{print}{FishStockValidation}(x, ...)
}
\arguments{
\item{x}{FishStockValidation object}

\item{...}{ignored}
}
\description{
Display a FishStockValidation object
}
