% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_lengths}
\alias{read_lengths}
\title{Import length frequency data}
\usage{
read_lengths(file)
}
\arguments{
\item{file}{file path}
}
\value{
standardized data.frame
}
\description{
The minimal expected data are:
}
\details{
- length

Optional:

- year
- month
- sex
- station
}
