% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{run_elefan_ga}
\alias{run_elefan_ga}
\title{ELEFAN by genetic algorithm (ELEFAN_GA)}
\usage{
run_elefan_ga(lfq, control = list(), ...)
}
\arguments{
\item{lfq}{Object of class \code{lfq} (TropFishR).}

\item{control}{List of arguments passed to \code{ELEFAN_GA()} : typically
\code{low_par}, \code{up_par}, \code{popSize}, \code{maxiter},
\code{seasonalised}.}

\item{...}{Additional arguments passed to \code{ELEFAN_GA()}.}
}
\value{
An object of class \code{FishStockELEFAN}.
}
\description{
Wrapper around \code{TropFishR::ELEFAN_GA()} : optimization of growth
parameters by genetic algorithm. The control arguments are filtered according
to the signature of the installed version of TropFishR, which ensures
compatibility across versions.
}
\examples{
\dontrun{
  fit <- run_elefan_ga(lfq, control = list(
    low_par = list(Linf = 40, K = 0.4, t_anchor = 0),
    up_par  = list(Linf = 55, K = 2.0, t_anchor = 1),
    popSize = 40, maxiter = 30))
}
}
\references{
Taylor, M.H., Mildenberger, T.K. (2017). Extending
  electronic length frequency analysis in R. \emph{Fisheries Management and
  Ecology} 24(4).
}
\seealso{
\code{\link{run_elefan_sa}}, \code{\link{compare_growth_models}}
}
