% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{run_elefan_sa}
\alias{run_elefan_sa}
\title{ELEFAN by simulated annealing (ELEFAN_SA)}
\usage{
run_elefan_sa(lfq, control = list(), ...)
}
\arguments{
\item{lfq}{Object of class \code{lfq} (TropFishR).}

\item{control}{List of arguments passed to \code{ELEFAN_SA()} : typically
\code{low_par}, \code{up_par}, \code{SA_time}, \code{SA_temp},
\code{seasonalised}.}

\item{...}{Additional arguments passed to \code{ELEFAN_SA()}.}
}
\value{
An object of class \code{FishStockELEFAN}.
}
\description{
Wrapper around \code{TropFishR::ELEFAN_SA()} : optimization of growth
parameters by simulated annealing.
}
\examples{
\dontrun{
  fit <- run_elefan_sa(lfq, control = list(
    low_par = list(Linf = 40, K = 0.4, t_anchor = 0),
    up_par  = list(Linf = 55, K = 2.0, t_anchor = 1)))
}
}
\references{
Taylor, M.H., Mildenberger, T.K. (2017). \emph{Fisheries
  Management and Ecology} 24(4).
}
\seealso{
\code{\link{run_elefan_ga}}, \code{\link{compare_growth_models}}
}
