% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{run_thompson_bell}
\alias{run_thompson_bell}
\title{Thompson-Bell yield prediction (TropFishR)}
\usage{
run_thompson_bell(param, FM_change = seq(0, 2, 0.1), Lc_change = NULL, ...)
}
\arguments{
\item{param}{List of TropFishR-compatible parameters (Linf, K, M, a, b,
plus composition/selectivity depending on version).}

\item{FM_change}{Vector of F multipliers to explore.}

\item{Lc_change}{Vector of sizes at first capture to explore
(optional, for a mesh-size x effort isopleth analysis).}

\item{...}{Arguments passed to \code{TropFishR::predict_mod()}.}
}
\value{
Raw output of \code{TropFishR::predict_mod()}.
}
\description{
Wrapper around \code{TropFishR::predict_mod(type = "ThompBell")} to
evaluate the effect of a change in effort (\code{FM_change}) and/or size
at first capture (\code{Lc_change}) on yield and biomass.
}
