% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{save_stock}
\alias{save_stock}
\title{Save FishStockData object}
\usage{
save_stock(stock, file = "data/processed/FishStockData.rds")
}
\arguments{
\item{stock}{\code{FishStockData} object.}

\item{file}{Path of the output RDS file.}
}
\value{
The path of the written file (invisible).
}
\description{
Save FishStockData object
}
\examples{
\dontrun{
  save_stock(stock, "data/processed/FishStockData.rds")
}
}
\seealso{
\code{\link{load_stock}}
}
