% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{summarize_import}
\alias{summarize_import}
\title{Summarize an imported dataset}
\usage{
summarize_import(data)
}
\arguments{
\item{data}{Imported \code{data.frame}.}
}
\value{
A summary \code{data.frame}.
}
\description{
Computes basic descriptive statistics of an imported dataset (number
of rows, columns, missing values).
}
\examples{
\dontrun{
  summarize_import(lengths)
}
}
