% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{summary.FishStockKScan}
\alias{summary.FishStockKScan}
\title{Summary of a K-Scan}
\usage{
\method{summary}{FishStockKScan}(object, ...)
}
\arguments{
\item{object}{Object \code{FishStockKScan}.}

\item{...}{Ignored.}
}
\description{
Summary of a K-Scan
}
