% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{summary.FishStockPowell}
\alias{summary.FishStockPowell}
\title{Summary of a FishStockPowell object}
\usage{
\method{summary}{FishStockPowell}(object, ...)
}
\arguments{
\item{object}{FishStockPowell object}

\item{...}{ignored}
}
\description{
Summary of a FishStockPowell object
}
