% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_catches}
\alias{total_catches}
\title{Annual total catches by group}
\format{
A data.frame with 55 rows and 4 variables:
\describe{
  \item{year}{year.}
  \item{Cymbium}{Catches of \emph{Cymbium} spp. (t).}
  \item{Octopus}{Catches of \emph{Octopus vulgaris} (t).}
  \item{Penaeus}{Catches of \emph{Penaeus notialis} (t).}
}
}
\source{
Fictitious simulated data (see \code{data-raw/simulate_datasets.R}).
}
\usage{
total_catches
}
\description{
Annual catch series (tonnes) for the three target groups.
}
\keyword{datasets}
